[REFACTOR-015] decompose: extract MembersTable, MembersFilterBar, and member mutation helpers from MembersTab
